Update publish command to webpack custom ds processors
