+++ packages/cli/src/commands/build.ts
@@ -4,45 +4,9 @@
 import {lstatSync, readFileSync} from 'fs';
 import path from 'path';
 import {Command, flags} from '@oclif/command';
-import webpack from 'webpack';
-import {merge} from 'webpack-merge';
+import cli from 'cli-ux';
+import {runWebpack} from '../controller/build-controller';
 import Validate from './validate';
-
-const getBaseConfig = (dir: string, outputPath: string, development?: boolean): webpack.Configuration => ({
-  target: 'node',
-  mode: development ? 'development' : 'production',
-  entry: path.join(dir, 'src/index.ts'),
-  devtool: development && 'inline-source-map',
-  module: {
-    rules: [
-      {
-        test: /\.tsx?$/,
-        exclude: /node_modules/,
-        loader: require.resolve('ts-loader'),
-        options: {
-          compilerOptions: {
-            declaration: false,
-          },
-        },
-      },
-      {
-        test: /\.ya?ml$/,
-        use: 'yaml-loader',
-      },
-    ],
-  },
-
-  resolve: {
-    extensions: ['.tsx', '.ts', '.js', '.json'],
-  },
-
-  output: {
-    path: path.dirname(outputPath),
-    filename: path.basename(outputPath),
-    libraryTarget: 'commonjs',
-    clean: true,
-  },
-});
 
 export default class Build extends Command {
   static description = 'Build this SubQuery project code';
@@ -73,32 +37,8 @@ export default class Build extends Command {
     const pjson = JSON.parse(readFileSync(path.join(directory, 'package.json')).toString());
     const outputPath = path.resolve(directory, pjson.main || 'dist/index.js');
 
-    const config = merge(
-      getBaseConfig(directory, outputPath, isDev)
-      // Can allow projects to override webpack config here
-    );
-
-    // Use webpack to build TS code and package into a single file
-    this.log(`Building code${isDev ? ' with development mode' : ''}`);
-    await new Promise((resolve, reject) => {
-      webpack(config).run((error, stats) => {
-        if (error) {
-          reject(error);
-          this.log(error.message);
-          return;
-        }
-
-        if (stats.hasErrors()) {
-          const info = stats.toJson();
-
-          reject(info.errors[0]);
-          this.log(info.errors[0].details);
-          return;
-        }
-
-        this.log('Finished building code');
-        resolve(true);
-      });
-    });
+    cli.action.start('Building and packing code');
+    await runWebpack(path.join(directory, 'src/index.ts'), outputPath, isDev, true);
+    cli.action.stop();
   }
 }
